Creating this project with MIT license
I'm normally partial to GPL licenses to avoid large corporation abuse, but this is designed to just be a dumb Dynamic DNS script to be run on a raspberry pi. Some small businesses (who don't want to pay for a dedicated IP) may want to use it, and I've been the IT guy for those type of companies before. Adding any more stress about licensing is the last thing they need. So, I'll start this project off with the very open MIT license.