Init commit
Initial commit. All these files are actually copied from a different private repository, with 82 commits. Originally I was just going to make that private repo public, but I realized there's too much personal information buried in the old commits. I've been updating every script in the project to remove any traces of personal information since then. But I realized now is a good time to make it public. It's kinda weird to upload an entire project in 1 commit to github. That history is useful. So while I'm still mid-way throught updating this project to it's final stages, I've completed the personal information purge, and now is a great time to start tracking update history in a new public repository.