Standardized indentation in README.md file.
The auto Table of Contents generator from Markdown-All-in-One vscode extension will always make an indentation 2 spaces. This works fine for that so we won't touch it. But it doesn't get formatted well in the rest of the document without 3 spaces or more. So for now on, all README.md indentation will be 4 spaces, no tabs, except in Table of Contents block that gets auto-generated.