Full test of setup.sh completed for v1. Many hard-coded vars were fixed to make relative references instead (including in rand.sh as well). Adding and removing cron jobs updated to always target the specific user's crontab file.
The _TEMPLATE params files have also been updated to include better instructions at the top.

Together all the changes in this commit should bring the repo into a state where most people should be able to get it up-and-running with no errors in about 5 minutes. Ready for an actual v1.0 release on github.